implemented comments from review
* removed the -i text in the help
* created a combined ExperimentConfig - put all the Configs in one file
* changed the Exception to ValueError and KeyError
* added a little text output for the console once the program is started
