+++ src/hivpy/experiment.py
@@ -1,44 +1,23 @@
-from dataclasses import dataclass
 from datetime import date, timedelta
-from .simulation import run_simulation, SimulationConfig
-import logging
+from .simulation import run_simulation
+from .config import SimulationConfig, OutputConfig
 import os
-
-LEVELS = {
-    'DEBUG': logging.DEBUG,
-    'INFO': logging.INFO,
-    'WARNING': logging.WARNING,
-    'ERROR': logging.ERROR,
-    'CRITICAL': logging.CRITICAL
-}
-
-@dataclass
-class OutputConfig:
-    output_dir: str
-    logfile: str
-    loglevel: str
-
-    def start_logging(self):
-        logging.basicConfig(filename=self.logfile, level=LEVELS[self.loglevel])
-        logging.info("starting experiment")
-
 
 
 def create_experiment(experiment_param):
-    start_date = date(1980,1,1)
-    end_date = date(2040, 12, 31)
-    interval = timedelta(days=30)
-    population_size = 1000
     try:
         start_date = date(int(experiment_param['START_YEAR']),1 ,1 )
         end_date = date(int(experiment_param['END_YEAR']),12, 31)
         population_size = int(experiment_param['POPULATION'])
         interval = timedelta(days = int(experiment_param['TIME_INTERVAL_DAYS']))
-    except RuntimeError as err:
+        return SimulationConfig(population_size, start_date, end_date, interval)
+    except ValueError as err:
         print('Error parsing the experiment parameters {}'.format(err))
+    except KeyError as kerr:
+        print('Error extracting values from the parameter set {}'.format(kerr))
+    return None
 
     # Dummy values for now
-    return SimulationConfig(population_size, start_date, end_date, interval)
 
 def create_output(output_param):
     outputdir = output_param['OUTPUT_DIRECTORY']
@@ -49,11 +28,11 @@ def create_output(output_param):
 
 
 
-def run_experiment(experiment_config, output_config):
+def run_experiment(experiment_config):
     """Run an entire experiment.
 
     An experiment can consist of one or more simulation runs,
     as well as processing steps after those are completed.
     """
-    output_config.start_logging()
-    result = run_simulation(experiment_config)
+    experiment_config.output_config.start_logging()
+    result = run_simulation(experiment_config.simulation_config)
